# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nour WhatsApp Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WhatsappStatusStatus =
  (typeof WhatsappStatusStatus)[keyof typeof WhatsappStatusStatus];

export const WhatsappStatusStatus = {
  connected: "connected",
  disconnected: "disconnected",
  connecting: "connecting",
  qr_ready: "qr_ready",
  pairing_ready: "pairing_ready",
} as const;

export interface WhatsappStatus {
  connected: boolean;
  phone?: string | null;
  name?: string | null;
  status: WhatsappStatusStatus;
  pairingCode?: string | null;
}

export interface WhatsappQr {
  qr?: string | null;
  status: string;
}

export interface WhatsappPairingBody {
  phone: string;
}

export interface WhatsappPairingResponse {
  success: boolean;
  pairingCode?: string | null;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Contact {
  id: number;
  phone: string;
  name?: string | null;
  firstSeen: string;
  lastSeen: string;
  messageCount: number;
  isBlocked: boolean;
  lastMessage?: string | null;
}

export interface ContactsListResponse {
  contacts: Contact[];
  total: number;
  page: number;
  totalPages: number;
}

export type MessageDirection =
  (typeof MessageDirection)[keyof typeof MessageDirection];

export const MessageDirection = {
  inbound: "inbound",
  outbound: "outbound",
} as const;

export interface Message {
  id: number;
  contactId: number;
  contactPhone: string;
  contactName?: string | null;
  content: string;
  direction: MessageDirection;
  aiModel?: string | null;
  timestamp: string;
}

export interface MessagesListResponse {
  messages: Message[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ContactDetail {
  contact: Contact;
  messages: Message[];
}

export interface SendMessageBody {
  phone: string;
  content: string;
}

export interface BroadcastMessageBody {
  content: string;
}

export interface BroadcastResult {
  success: boolean;
  sent: number;
  failed: number;
}

export interface BlockContactBody {
  blocked: boolean;
}

export interface AiModelOption {
  id: string;
  name: string;
  description: string;
}

export interface AiModelsResponse {
  models: AiModelOption[];
}

export type SettingsAiModel =
  (typeof SettingsAiModel)[keyof typeof SettingsAiModel];

export const SettingsAiModel = {
  gemini: "gemini",
  groq: "groq",
} as const;

export interface Settings {
  ownerName?: string;
  ownerEmail?: string;
  ownerPhone?: string;
  adminPhone?: string;
  projectName?: string;
  projectDescription?: string;
  projectLink?: string;
  demoVideoUrl?: string;
  geminiApiKey?: string;
  geminiApiKey2?: string;
  geminiApiKey3?: string;
  geminiApiKey4?: string;
  geminiApiKey5?: string;
  geminiApiKey6?: string;
  geminiModel?: string;
  groqApiKey?: string;
  groqModel?: string;
  aiModel: SettingsAiModel;
  agentPersonality?: string;
  autoReply: boolean;
  maintenanceMode?: boolean;
  maintenanceMessage?: string;
  firebaseProjectId?: string;
  firebaseClientEmail?: string;
  firebasePrivateKey?: string;
  firebaseBucket?: string;
}

export interface Stats {
  totalContacts: number;
  totalMessages: number;
  messagesSent: number;
  messagesReceived: number;
  blockedContacts: number;
  todayMessages: number;
  activeContacts: number;
}

export interface ActivityPoint {
  date: string;
  inbound: number;
  outbound: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type SystemLogLevel =
  (typeof SystemLogLevel)[keyof typeof SystemLogLevel];

export const SystemLogLevel = {
  info: "info",
  warn: "warn",
  error: "error",
  success: "success",
} as const;

export type SystemLogCategory =
  (typeof SystemLogCategory)[keyof typeof SystemLogCategory];

export const SystemLogCategory = {
  whatsapp: "whatsapp",
  ai: "ai",
  system: "system",
} as const;

export interface SystemLog {
  id: number;
  level: SystemLogLevel;
  category: SystemLogCategory;
  event: string;
  message: string;
  details?: string | null;
  timestamp: string;
}

export interface SystemLogsResponse {
  logs: SystemLog[];
  total: number;
}

export interface ClearDataResponse {
  success: boolean;
  deleted: number;
  message: string;
}

export type GetContactsParams = {
  search?: string;
  page?: number;
  limit?: number;
};

export type GetMessagesParams = {
  contactId?: number;
  direction?: GetMessagesDirection;
  page?: number;
  limit?: number;
};

export type GetMessagesDirection =
  (typeof GetMessagesDirection)[keyof typeof GetMessagesDirection];

export const GetMessagesDirection = {
  inbound: "inbound",
  outbound: "outbound",
} as const;

export type ClearMessagesParams = {
  contactId?: number;
};

export type GetGeminiModelsParams = {
  key: string;
};

export type GetGroqModelsParams = {
  key: string;
};

export type GetSystemLogsParams = {
  limit?: number;
  offset?: number;
  category?: string;
  level?: string;
};