# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nour WhatsApp Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    post:
      operationId: updateSettings
      tags: [settings]
      summary: Update settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /settings/models/gemini:
    get:
      operationId: getGeminiModels
      tags: [settings]
      summary: Fetch available Gemini models for a given API key
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of available models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModelsResponse"
        "400":
          description: Invalid or missing key
  /settings/models/groq:
    get:
      operationId: getGroqModels
      tags: [settings]
      summary: Fetch available Groq models for a given API key
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of available models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModelsResponse"
        "400":
          description: Invalid or missing key
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    AiModelOption:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
      required: [id, name, description]
    AiModelsResponse:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/AiModelOption"
      required: [models]
    Settings:
      type: object
      properties:
        ownerName:
          type: string
        ownerEmail:
          type: string
        ownerPhone:
          type: string
        adminPhone:
          type: string
        projectName:
          type: string
        projectDescription:
          type: string
        projectLink:
          type: string
        demoVideoUrl:
          type: string
        geminiApiKey:
          type: string
        geminiApiKey2:
          type: string
        geminiApiKey3:
          type: string
        geminiApiKey4:
          type: string
        geminiApiKey5:
          type: string
        geminiApiKey6:
          type: string
        geminiModel:
          type: string
        groqApiKey:
          type: string
        groqModel:
          type: string
        aiModel:
          type: string
          enum: [gemini, groq]
        agentPersonality:
          type: string
        autoReply:
          type: boolean
        maintenanceMode:
          type: boolean
        maintenanceMessage:
          type: string
        firebaseProjectId:
          type: string
        firebaseClientEmail:
          type: string
        firebasePrivateKey:
          type: string
        firebaseBucket:
          type: string
      required: [aiModel, autoReply]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...