# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nour WhatsApp Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsAiModel } from "./settingsAiModel";

export interface Settings {
  ownerName?: string;
  ownerEmail?: string;
  ownerPhone?: string;
  adminPhone?: string;
  projectName?: string;
  projectDescription?: string;
  projectLink?: string;
  demoVideoUrl?: string;
  geminiApiKey?: string;
  geminiApiKey2?: string;
  geminiApiKey3?: string;
  geminiApiKey4?: string;
  geminiApiKey5?: string;
  geminiApiKey6?: string;
  geminiModel?: string;
  groqApiKey?: string;
  groqModel?: string;
  aiModel: SettingsAiModel;
  agentPersonality?: string;
  autoReply: boolean;
  maintenanceMode?: boolean;
  maintenanceMessage?: string;
  firebaseProjectId?: string;
  firebaseClientEmail?: string;
  firebasePrivateKey?: string;
  firebaseBucket?: string;
}